Float casting for temperatures

+++ watch-faces/sensor/temperature_display_face.c
@@ -42,7 +42,7 @@ void temperature_display_face_setup(uint8_t watch_face_index, void ** context_pt
     (void) watch_face_index;
     (void) context_ptr;
     // if temperature is invalid, we don't have a temperature sensor which means we shouldn't be here.
-    if (movement_get_temperature() == 0xFFFFFFFF) skip = true;
+    if (movement_get_temperature() == (float)0xFFFFFFFF) skip = true;
 }
 
 void temperature_display_face_activate(void *context) {
